requirements.txt
absl-py==2.3.1
altgraph==0.17.4
astunparse==1.6.3
async-timeout==5.0.1
bleak==1.1.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.0
fsspec==2025.9.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.75.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mne==1.10.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
python-picard==0.8
pytz==2025.2
pywin32-ctypes==0.2.3
QtAwesome==1.4.0
QtPy==2.4.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.9.0+cu126
torchvision==0.24.0+cu126
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Bluetooth.Advertisement==3.2.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Storage.Streams==3.2.1
wrapt==1.17.3
